requirements.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.1.4
numpy==1.26.0
google-generativeai==0.3.2
openai==1.10.0
websocket-client==1.7.0
python-dotenv==1.0.0
requests==2.31.0
pyngrok==7.0.0
python-binance==1.0.19
ta==0.11.0
aiohttp==3.9.1
scipy==1.11.4
anthropic==0.8.1
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
redis==5.0.1
twilio==8.10.0
python-telegram-bot==20.7
pyqrcode==1.2.1
pypng==0.20220715.0
pyperclip==1.8.2
